pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aml-investigator-env"
version = "1.0.0"
description = "AML Investigator OpenEnv Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic",
    "openai",
    "openenv-core>=0.2.0",
    "fastapi",
    "uvicorn",
    "python-dotenv"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["env", "models", "inference", "app", "dashboard", "agent", "tasks"]
packages = ["server"]